Big performance improvements and UI improvements
Now the bars of players will stay in the same position until they
speak again. This helps again rapidly disappearing and appearing
bars, since it will not disturb the position of the bars of
actively talking players.

The whole voice chat file has also been replaced now, cleaning up
the code, not using slow functions, and a GetIsVisible method
to help ns2 opti.

All bars are also now preallocated, the number of which depends on the size of your screen.