FINALLY fixed the flickering name bug, and code refactoring
This was quite an annoying bug, and the cause was finally found:
The speakers were not relevant for the listener, and therefore the
listener wasn't always having them set as VoiceChannel.Entity, and it's
also why you couldn't here them, since their position was not available
to the listener!

This is now fixed by having a relevancy check on the client, so that
irrelevant speakers are not considered in the algorithm.

In addition to this, all use of the scoreboard API was removed, now the
leaner PlayerInfoEntity system is used, which the scoreboard API is based
on actually.

This does bring some performance benefits, but most importantly, we gain
assurance that weird behavior won't happen, since player records aren't removed
immediately. Player info entities are always removed immediately when the client
disconnects, so this is no problem.

A minor name convention change was also done for the voice chat bars.