Fix bars not always showing and disallow specs from talking in a round in progress
Since voice chat bars now rely on the player of the player info being relevant,
voice chat bars were not shown for spectators who were spectating others.

The reason this even works is because the engine knows who is spectating who,
so it can get the origin from the spectated player, if they are relevant,
even though the spectator itself is irrelevant.

If the spectated player is NOT relevant, we encounter the old bug again,
but this is most likely a very rare case, so I'd rather not spend time fixing this bug.