Allow listening to team-only local voice chat as spectator, and fix colors when speaking team-only as spectator
There's a weird mismatch between the team number of a spectator
and the team number of the correponding player info, where the
player info will contain the team number of the spectator itself,
but the spectator entity itself will contain the team number
of the followed player, and not the actual team number.

This is quite odd behavior and was the cause of a logic bug.